requirements.txt
aio-pika==9.5.3
aiofiles==24.1.0
aioflask==0.4.0
aioredis==2.0.1
aiormq==6.8.1
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.2.post1
asgiref==3.8.1
async==0.6.2
async-timeout==5.0.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Django==4.2.23
djangorestframework==3.16.0
dnspython==2.7.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
fastapi==0.115.5
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
greenlet==3.1.1
greenletio==0.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
motor==3.6.0
multidict==6.1.0
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pamqp==3.3.0
pika==1.3.2
pluggy==1.5.0
priority==2.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
pymongo==3.4.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
Quart==0.19.9
Quart-Bcrypt==0.0.5
quart-cors==0.7.0
redis==5.2.0
requests==2.32.4
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
taskgroup==0.0.0a4
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
vboxapi==1.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
